+++ src/client/node_serial_impl.ts
@@ -1,6 +1,6 @@
 import { SerialPort } from "serialport";
 import { ConnectEvent, DisconnectEvent, RawPacketSentEvent } from "../events";
-import { ConnectionInfo, NiimbotAbstractClient } from ".";
+import { ConnectionInfo, NiimbotAbstractClient, NiimbotClientType } from ".";
 import { ConnectResult } from "../packets";
 import { Utils } from "../utils";
 
@@ -160,5 +160,9 @@ export class NiimbotNodeSerialClient extends NiimbotAbstractClient {
         address: p.path,
       };
     });
+  }
+
+  public override getType(): NiimbotClientType {
+    return "node-serial";
   }
 }