+++ src/client/serial_impl.ts
@@ -1,5 +1,5 @@
 import { ConnectEvent, DisconnectEvent, RawPacketSentEvent } from "../events";
-import { ConnectionInfo, NiimbotAbstractClient } from ".";
+import { ConnectionInfo, NiimbotAbstractClient, NiimbotClientType } from ".";
 import { ConnectResult } from "../packets";
 import { Utils } from "../utils";
 
@@ -121,5 +121,9 @@ export class NiimbotSerialClient extends NiimbotAbstractClient {
     } else {
       await this.mutex.runExclusive(send);
     }
+  }
+
+  public override getType(): NiimbotClientType {
+    return "serial";
   }
 }